wrap up reshuffle of fig builder / data processor / field manipulator utility functions
