+++ app/fig_builder/plotly_gantt.py
@@ -5,7 +5,6 @@ import plotly.figure_factory as ff
 import plotly.graph_objects as go
 import random
 
-import app.fig_builder.fig_helper as fh
 import app.fig_meta.color_meta as cm
 import app.fig_meta.gantt_helper as gh
 from app.show_metadata import show_metadata
